Support previewMaxFrameRate parameter of Options API PUT.

+++ dConnectDevicePlugin/dConnectDeviceHost/app/src/main/java/org/deviceconnect/android/deviceplugin/host/profile/RequestParam.java
@@ -92,12 +92,12 @@ class RequestParam {
         throw new NumberFormatException();
     }
 
-    private int parseDouble(Object obj) {
+    private double parseDouble(Object obj) {
         if (obj instanceof Integer) {
             return (Integer) obj;
         }
         if (obj instanceof String) {
-            return Integer.parseInt((String) obj);
+            return Double.parseDouble((String) obj);
         }
         throw new NumberFormatException();
     }
